notification-service/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic-settings==2.3.0
python-dotenv==1.0.1
redis==5.0.1
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-cov==5.0.0
httpx==0.27.0
pytest-mock==3.14.0
slowapi==0.1.9

# Bundle F: Enterprise Features
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Bundle G: DevOps & Performance
prometheus-client==0.20.0
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-exporter-jaeger==1.21.0